+++ lib/eve_online/esi/dogma_attributes.rb
@@ -1,0 +1,17 @@
+module EveOnline
+  module ESI
+    class DogmaAttributes < Base
+      API_ENDPOINT = 'https://esi.tech.ccp.is/v1/dogma/attributes/?datasource=tranquility'.freeze
+
+      def attributes
+        response
+      end
+
+      def scope; end
+
+      def url
+        API_ENDPOINT
+      end
+    end
+  end
+end